requirements.txt
pandas==1.5.3
streamlit
folium
streamlit_folium
geopandas
plotly
Pillow